% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{plot.slots}
\alias{plot.slots}
\title{Plot method for Slots}
\usage{
\method{plot}{slots}(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot method for Slots
}
