% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(symbols, version = 0)
}
\arguments{
\item{symbols}{A dataframe of plays from play()}

\item{version}{Which version of score() to use}
}
\value{
A score for an oucome from play()
}
\description{
Score output from play()
}
